requirements.txt
flask==3.0.3
python-dotenv==1.0.1

langchain==0.2.16
langchain-core==0.2.38
langchain-community==0.2.16
langchain-google-genai==1.0.10
langchain-pinecone==0.1.3
langchain-text-splitters==0.2.4

pinecone-client==5.0.1
sentence-transformers==2.7.0
huggingface-hub==0.24.6
pypdf==4.3.1

